Refactor Name#primer
Because RuboCop noted an AbcSize offense.
It was also convenient to extract the refactored method to its own file.
